+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Use `package:clock` and `package:fake_async` across the board.
 * Optimize the implementation of 'ExpiryCache'.
 * Fix `maxAge` of the `Stream.buffer` operator.
+* Add `Fraction.sign` and `Complex.sign`.
 
 ## 2.8.0